requirements.txt
numpy
Pillow
progressbar2
termcolor
stop-words
